requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
PyJWT>=2.8,<3.0
django-environ>=0.11,<1.0
